+++ src/main/scala/au/com/agl/arc/extract/DelimitedExtract.scala
@@ -41,6 +41,15 @@ object DelimitedExtract {
       .map("stage", stageDetail)      
       .log()    
     
+    // try to get the schema
+    val optionSchema = try {
+      ExtractUtils.GetSchema(extract.cols)(spark)
+    } catch {
+      case e: Exception => throw new Exception(e) with DetailException {
+        override val detail = stageDetail          
+      }      
+    }        
+
     val df = try {
       extract.input match {
         case Right(glob) =>
@@ -62,25 +71,21 @@ object DelimitedExtract {
     }
 
     // if incoming dataset has 0 columns then create empty dataset with correct schema
-    val emptyDataframeHandlerDF = if (df.schema.length == 0) {
-      val schema = extract.cols match {
-        case Nil => throw new Exception(s"DelimitedExtract has produced 0 columns and no schema has been provided to create an empty dataframe.") with DetailException {
-          override val detail = stageDetail          
-        }
-        case cols => Extract.toStructType(cols)
-      }
-      spark.createDataFrame(spark.sparkContext.emptyRDD[Row], schema)
-    } else {
-      df
-    }        
+    val emptyDataframeHandlerDF = try {
+      ExtractUtils.EmptyDataFrameHandler(df, optionSchema)(spark)
+    } catch {
+      case e: Exception => throw new Exception(e.getMessage) with DetailException {
+        override val detail = stageDetail          
+      }      
+    }      
 
-    // add source data including index
-    val sourceEnrichedDF = ExtractUtils.addSourceMetadata(emptyDataframeHandlerDF, contiguousIndex)
+    // add internal columns data _filename, _index
+    val sourceEnrichedDF = ExtractUtils.AddInternalColumns(emptyDataframeHandlerDF, contiguousIndex)
 
     // set column metadata if exists
-    val enrichedDF = extract.cols match {
-      case Nil => sourceEnrichedDF
-      case cols => MetadataUtils.setMetadata(sourceEnrichedDF, Extract.toStructType(cols))
+    val enrichedDF = optionSchema match {
+        case Some(schema) => MetadataUtils.setMetadata(sourceEnrichedDF, schema)
+        case None => sourceEnrichedDF   
     }
          
     // repartition to distribute rows evenly