+++ src/test/scala/au/com/agl/arc/extract/JSONExtractSuite.scala
@@ -76,7 +76,7 @@ class JSONExtractSuite extends FunSuite with BeforeAndAfter {
     val extractDataset = extract.JSONExtract.extract(
       JSONExtract(
         name=outputView,
-        cols=cols.right.getOrElse(Nil),
+        cols=Right(cols.right.getOrElse(Nil)),
         outputView=outputView,
         input=Right(targetFileGlob),
         settings=new JSON(multiLine=false),
@@ -121,7 +121,7 @@ class JSONExtractSuite extends FunSuite with BeforeAndAfter {
     extract.JSONExtract.extract(
       JSONExtract(
         name=outputView,
-        cols=Nil,
+        cols=Right(Nil),
         outputView=outputView,
         input=Right(targetFile),
         settings=new JSON(multiLine=false),
@@ -139,7 +139,7 @@ class JSONExtractSuite extends FunSuite with BeforeAndAfter {
     extract.JSONExtract.extract(
       JSONExtract(
         name=outputView,
-        cols=Nil,
+        cols=Right(Nil),
         outputView=outputView,
         input=Right(targetFile),
         settings=new JSON(),
@@ -178,7 +178,7 @@ class JSONExtractSuite extends FunSuite with BeforeAndAfter {
       val extractDataset = extract.JSONExtract.extract(
         JSONExtract(
           name=outputView,
-          cols=Nil,
+          cols=Right(Nil),
           outputView=outputView,
           input=Right(emptyWildcardDirectory),
           settings=new JSON(multiLine=false),
@@ -192,14 +192,14 @@ class JSONExtractSuite extends FunSuite with BeforeAndAfter {
       )
     }
 
-    assert(thrown0.getMessage === "JSONExtract has produced 0 columns and no schema has been provided to create an empty dataframe.")
+    assert(thrown0.getMessage === "Extract has produced 0 columns and no schema has been provided to create an empty dataframe.")
 
     // try without providing column metadata
     val thrown1 = intercept[Exception with DetailException] {
       val extractDataset = extract.JSONExtract.extract(
         JSONExtract(
           name=outputView,
-          cols=Nil,
+          cols=Right(Nil),
           outputView=outputView,
           input=Right(emptyDirectory),
           settings=new JSON(multiLine=false),
@@ -213,13 +213,13 @@ class JSONExtractSuite extends FunSuite with BeforeAndAfter {
       )
     }
 
-    assert(thrown1.getMessage === "JSONExtract has produced 0 columns and no schema has been provided to create an empty dataframe.")
+    assert(thrown1.getMessage === "Extract has produced 0 columns and no schema has been provided to create an empty dataframe.")
     
     // try with column
     val extractDataset = extract.JSONExtract.extract(
       JSONExtract(
         name=outputView,
-        cols=cols,
+        cols=Right(cols),
         outputView=outputView,
         input=Right(emptyDirectory),
         settings=new JSON(multiLine=false),
@@ -250,7 +250,7 @@ class JSONExtractSuite extends FunSuite with BeforeAndAfter {
     val actual0 = extract.JSONExtract.extract(
       JSONExtract(
         name=outputView,
-        cols=Nil,
+        cols=Right(Nil),
         outputView=outputView,
         input=Right(multiLineMatcher),
         settings=new JSON(multiLine=false),
@@ -266,7 +266,7 @@ class JSONExtractSuite extends FunSuite with BeforeAndAfter {
     val actual1 = extract.JSONExtract.extract(
       JSONExtract(
         name=outputView,
-        cols=Nil,
+        cols=Right(Nil),
         outputView=outputView,
         input=Right(multiLineMatcher),
         settings=new JSON(multiLine=true),
@@ -321,7 +321,7 @@ test("JSONExtract: Input Schema") {
     val extractDataset = extract.JSONExtract.extract(
       JSONExtract(
         name=outputView,
-        cols=cols,
+        cols=Right(cols),
         outputView=outputView,
         input=Right(targetFile),
         settings=new JSON(multiLine=false),