+++ src/test/scala/au/com/agl/arc/extract/ParquetExtractSuite.scala
@@ -65,7 +65,7 @@ class ParquetExtractSuite extends FunSuite with BeforeAndAfter {
     val extractDataset = extract.ParquetExtract.extract(
       ParquetExtract(
         name=outputView,
-        cols=cols.right.getOrElse(Nil),
+        cols=Right(cols.right.getOrElse(Nil)),
         outputView=outputView,
         input=targetFileGlob,
         authentication=None,
@@ -100,7 +100,7 @@ class ParquetExtractSuite extends FunSuite with BeforeAndAfter {
     extract.ParquetExtract.extract(
       ParquetExtract(
         name=outputView,
-        cols=Nil,
+        cols=Right(Nil),
         outputView=outputView,
         input=targetFile,
         authentication=None,
@@ -117,7 +117,7 @@ class ParquetExtractSuite extends FunSuite with BeforeAndAfter {
     extract.ParquetExtract.extract(
       ParquetExtract(
         name=outputView,
-        cols=Nil,
+        cols=Right(Nil),
         outputView=outputView,
         input=targetFile,
         authentication=None,
@@ -155,7 +155,7 @@ class ParquetExtractSuite extends FunSuite with BeforeAndAfter {
       val extractDataset = extract.ParquetExtract.extract(
         ParquetExtract(
           name=outputView,
-          cols=Nil,
+          cols=Right(Nil),
           outputView=outputView,
           input=emptyWildcardDirectory,
           authentication=None,
@@ -167,14 +167,14 @@ class ParquetExtractSuite extends FunSuite with BeforeAndAfter {
         )
       )
     }
-    assert(thrown0.getMessage === "ParquetExtract has produced 0 columns and no schema has been provided to create an empty dataframe.")
+    assert(thrown0.getMessage === "Extract has produced 0 columns and no schema has been provided to create an empty dataframe.")
     
     // try without providing column metadata
     val thrown1 = intercept[Exception with DetailException] {
       val extractDataset = extract.ParquetExtract.extract(
         ParquetExtract(
           name=outputView,
-          cols=Nil,
+          cols=Right(Nil),
           outputView=outputView,
           input=emptyDirectory,
           authentication=None,
@@ -186,13 +186,13 @@ class ParquetExtractSuite extends FunSuite with BeforeAndAfter {
         )
       )
     }
-    assert(thrown1.getMessage === "ParquetExtract has produced 0 columns and no schema has been provided to create an empty dataframe.")
+    assert(thrown1.getMessage === "Extract has produced 0 columns and no schema has been provided to create an empty dataframe.")
     
     // try with column
     val extractDataset = extract.ParquetExtract.extract(
       ParquetExtract(
         name=outputView,
-        cols=cols,
+        cols=Right(cols),
         outputView=outputView,
         input=emptyDirectory,
         authentication=None,